RAC-136 refactor : jwt 세팅 변경에 따른 수정
